+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 ### Features Added
 
 - Support for launching smaller service level MCP servers. https://github.com/Azure/azure-mcp/pull/324
+- Added support for Key Vault Secrets. https://github.com/Azure/azure-mcp/pull/173
 
 ### Bugs Fixed